MainMenuShop-Front/src/app/pages/login/login.component.ts
import { Router } from '@angular/router';
import { LoginService } from './../../services/login.service';
import { MatSnackBar } from '@angular/material/snack-bar';
import { Component, OnInit } from '@angular/core';

@Component({
  selector: 'app-login',
  templateUrl: './login.component.html',
  styleUrls: ['./login.component.css']
})
export class LoginComponent implements OnInit {

  loginData = {
    "email": "",
    "password": ""
  }

  constructor(private snack:MatSnackBar,private loginService:LoginService,private router:Router) { }

  ngOnInit(): void {
  }

  formSubmit(){
    console.log(this.loginData);

    this.loginService.generateToken(this.loginData).subscribe(
      (data:any) => {

        this.loginService.loginUser(data.token);
        this.router.navigate(['/UserDashboardComponent']);

      },(error) => {
        this.snack.open('Detalles inválidos , vuelva a intentar !!','Aceptar',{
          duration:3000
        })
      }
    )
  }
}
